Bringing in Dev work

+++ app/controllers/admin/volunteers.py
@@ -10,13 +10,12 @@ from app.models.eventParticipant import EventParticipant
 from app.logic.searchUsers import searchUsers
 from app.logic.volunteers import updateEventParticipants, addVolunteerToEventRsvp, getEventLengthInHours, addUserBackgroundCheck, setProgramManager
 from app.logic.participants import trainedParticipants, getEventParticipants, addPersonToEvent
-from app.logic.events import getPreviousRecurringEventData
+from app.logic.events import getPreviousRecurringEventData, getEventRsvpCountsForTerm
 from app.models.eventRsvp import EventRsvp
 from app.models.backgroundCheck import BackgroundCheck
 from app.models.programManager import ProgramManager
-from app.logic.adminLogs import createLog
+from app.logic.createLogs import createAdminLog, createRsvpLog
 from app.logic.users import getBannedUsers, isBannedFromEvent
-
 
 
 @admin_bp.route('/searchVolunteers/<query>', methods = ['GET'])
@@ -25,14 +24,28 @@ def getVolunteers(query):
 
     return json.dumps(searchUsers(query))
 
-@admin_bp.route('/eventsList/<eventID>/track_volunteers', methods=['GET'])
-def trackVolunteersPage(eventID):
+@admin_bp.route('/event/<eventID>/track_volunteers', methods=['POST'])
+def updateVolunteerTable(eventID):
     try:
         event = Event.get_by_id(eventID)
     except DoesNotExist as e:
         print(f"No event found for {eventID}")
         abort(404)
 
+    volunteerUpdated = updateEventParticipants(request.form)
+    if volunteerUpdated:
+        flash("Volunteer table succesfully updated", "success")
+    else:
+        flash("Error adding volunteer", "danger")
+    return redirect(url_for("admin.trackVolunteersPage", eventID=eventID))
+
+@admin_bp.route('/event/<eventID>/track_volunteers', methods=['GET'])
+def trackVolunteersPage(eventID):
+    try:
+        event = Event.get_by_id(eventID)
+    except DoesNotExist as e:
+        print(f"No event found for {eventID}", e)
+        abort(404)
     eventData = model_to_dict(event, recurse=False)
     eventData["program"] = event.singleProgram
     trainedParticipantsList = trainedParticipants(event.singleProgram, g.current_term)
@@ -51,40 +64,61 @@ def trackVolunteersPage(eventID):
         if volunteer.user not in volunteerUser:
             eventVolunteerData.append(volunteer)
             volunteerUser.append(volunteer.user)
+    eventWaitlistData = [volunteer for volunteer in eventVolunteerData if volunteer.rsvpWaitlist]
     eventLengthInHours = getEventLengthInHours(event.timeStart, event.timeEnd, event.startDate)
 
     recurringEventID = event.recurringId # query Event Table to get recurringId using Event ID.
     recurringEventStartDate = event.startDate
     recurringVolunteers = getPreviousRecurringEventData(recurringEventID)
 
-    return render_template("/events/trackVolunteers.html",
-        eventData = eventData,
-        eventVolunteerData = eventVolunteerData,
-        eventParticipants = eventParticipants,
-        eventLength = eventLengthInHours,
-        event = event,
-        recurringEventID = recurringEventID,
-        recurringEventStartDate = recurringEventStartDate,
-        recurringVolunteers = recurringVolunteers,
-        bannedUsers = bannedUsers,
-        trainedParticipantsList = trainedParticipantsList)
+    currentRsvpAmount = getEventRsvpCountsForTerm(g.current_term)
 
-@admin_bp.route('/eventsList/<eventID>/track_volunteers', methods=['POST'])
-def updateVolunteerTable(eventID):
+    return render_template("/events/trackVolunteers.html",
+                            eventData = eventData,
+                            eventVolunteerData = eventVolunteerData,
+                            eventParticipants = eventParticipants,
+                            eventLength = eventLengthInHours,
+                            event = event,
+                            recurringEventID = recurringEventID,
+                            recurringEventStartDate = recurringEventStartDate,
+                            recurringVolunteers = recurringVolunteers,
+                            bannedUsers = bannedUsers,
+                            trainedParticipantsList = trainedParticipantsList,
+                            eventWaitlistData = eventWaitlistData,
+                            currentRsvpAmount = currentRsvpAmount)
+
+
+
+@admin_bp.route('/event/<eventID>/dietary_restrictions', methods=['GET'])
+def dietaryRestrictionsPage(eventID):
     try:
         event = Event.get_by_id(eventID)
     except DoesNotExist as e:
-        print(f"No event found for {eventID}")
+        print(f"No event found for {eventID}", e)
         abort(404)
 
-    program = event.singleProgram
 
-    volunteerUpdated = updateEventParticipants(request.form)
-    if volunteerUpdated:
-        flash("Volunteer table succesfully updated", "success")
-    else:
-        flash("Error adding volunteer", "danger")
-    return redirect(url_for("admin.trackVolunteersPage", eventID=eventID))
+    if not (g.current_user.isCeltsAdmin or (g.current_user.isCeltsStudentStaff and g.current_user.isProgramManagerForEvent(event))):
+        abort(403)
+
+    eventRsvpData = list(EventRsvp.select().where(EventRsvp.event==event).order_by(EventRsvp.rsvpTime))
+    eventParticipantData = list(EventParticipant.select().where(EventParticipant.event==event))
+    participantsAndRsvp = (eventParticipantData + eventRsvpData)
+
+    #get unique list of users for each category waitlist/notwaitlist
+    volunteerUser = list(set([obj.user for obj in participantsAndRsvp if not obj.rsvpWaitlist and obj.user.dietRestriction]))
+    waitlistUser = list(set([obj.user for obj in participantsAndRsvp if obj.rsvpWaitlist and obj.user.dietRestriction]))
+
+
+    eventData = model_to_dict(event, recurse=False)
+    eventData["program"] = event.singleProgram
+
+
+    return render_template("/events/dietaryRestrictions.html",
+                            volunteerUser = volunteerUser,
+                            waitlistUser = waitlistUser,
+                            event = event,
+                            eventData = eventData)
 
 
 @admin_bp.route('/addVolunteersToEvent/<eventId>', methods = ['POST'])
@@ -112,15 +146,36 @@ def addVolunteer(eventId):
 
     return redirect(url_for('admin.trackVolunteersPage', eventID = eventId))
 
+@admin_bp.route('/rsvpFromWaitlist/<username>/<eventId>', methods = ['POST'])
+def rsvpFromWaitlist(username, eventId):
+    event = Event.get_by_id(eventId)
+    isProgramManager = g.current_user.isProgramManagerFor(event.singleProgram)
+    if g.current_user.isCeltsAdmin or (g.current_user.isCeltsStudentStaff and isProgramManager): 
+        waitlistUsers = EventRsvp.select(EventRsvp, User).join(User).where(EventRsvp.user == username, EventRsvp.event==eventId).execute()
+        if (waitlistUsers):
+            createRsvpLog(event.id, f"Moved {waitlistUsers[0].user.fullName} from waitlist to RSVP.")
+            (EventRsvp.update(rsvpWaitlist = False).where(EventRsvp.event_id == eventId, EventRsvp.user_id == username)).execute()
+    return ""
+
 @admin_bp.route('/addVolunteersToEvent/<username>/<eventId>/isBanned', methods = ['GET'])
 def isVolunteerBanned(username, eventId):
     return {"banned":1} if isBannedFromEvent(username, eventId) else {"banned":0}
 
-@admin_bp.route('/removeVolunteerFromEvent/<user>/<eventID>', methods = ['POST'])
-def removeVolunteerFromEvent(user, eventID):
-    (EventParticipant.delete().where(EventParticipant.user==user, EventParticipant.event==eventID)).execute()
-    (EventRsvp.delete().where(EventRsvp.user==user, EventRsvp.event==eventID)).execute()
-    flash("Volunteer successfully removed", "success")
+@admin_bp.route('/removeVolunteerFromEvent', methods = ['POST'])
+def removeVolunteerFromEvent():
+    user = request.form.get('username')
+    eventID = request.form.get('eventId')
+    if g.current_user.isAdmin:
+        userInRsvpTable = EventRsvp.select(EventRsvp, User).join(User).where(EventRsvp.user == username, EventRsvp.event==eventID).execute()
+        if (userInRsvpTable):
+            rsvpUser = userInRsvpTable[0]
+            if rsvpUser.rsvpWaitlist:
+                createRsvpLog(eventID, f"Removed {rsvpUser.user.fullName} from waitlist.")
+            else:
+                createRsvpLog(eventID, f"Removed {rsvpUser.user.fullName} from RSVP list.")
+        (EventParticipant.delete().where(EventParticipant.user==user, EventParticipant.event==eventID)).execute()
+        (EventRsvp.delete().where(EventRsvp.user==user, EventRsvp.event==eventID)).execute()
+        flash("Volunteer successfully removed", "success")
     return ""
 
 @admin_bp.route('/addBackgroundCheck', methods = ['POST'])
@@ -149,7 +204,7 @@ def updateProgramManager():
         username = User.get(User.username == data["user_name"])
         program = Program.get_by_id(data['program_id'])
         setProgramManager(data["user_name"], data["program_id"], data["action"])
-        createLog(f'{username.firstName} has been {data["action"]}ed as a Program Manager for {program.programName}')
+        createAdminLog(f'{username.firstName} has been {data["action"]}ed as a Program Manager for {program.programName}')
         return ""
     else:
         abort(403)
